Back to Windows for the first time this year to see how easy it is to get recent Python embedding to work. Also using TortoiseSVN for the first time. Both are doing well. However, Python crashes currently in new client spec. So do no sync! This may be Python's link pragma which pulls in a non-existent lib in debug builds. Or it might be the way ODE got built.
Also rearranged the way 3rd party libs are pulled in. Should make it easier to put everything together from scratch next time. 

git-svn-id: https://crag.googlecode.com/svn/trunk@289 03cb304c-929a-b8d2-5531-aad0e9782889

+++ src/crag/gfx/Renderer.cpp
@@ -251,7 +251,7 @@ TimeType gfx::Renderer::Render(Scene & scene, bool enable_vsync)
 	if (frame_count == 60) 
 	{
 		frame_count_reset_time = frame_time;
-		average_frame_time = static_cast<double>(fps_delta) / frame_count;
+		average_frame_time = static_cast<float>(fps_delta) / frame_count;
 		frame_count = 0;
 	}
